requirements.txt
Flask>=3.0
openai>=1.25
ddtrace>=2.0
python-dotenv>=1.0
gunicorn>=22.0